Changed TrafficDialPercentage to number json type
TrafficDialPercentage in the AGA code base is a float not an integer.  This updates the type to a float.  Also, it was realized we were not actually setting traffic dial percentage at all in the tests, so we were always getting the default value of 100, even if a customer specified a value.  We no abide by what the customer specified.

+++ aws-globalaccelerator-endpointgroup/src/main/kotlin/software/amazon/globalaccelerator/endpointgroup/ReadHandler.kt
@@ -44,7 +44,7 @@ class ReadHandler : BaseHandler<CallbackContext>() {
                 this.healthCheckPort = endpointGroup.healthCheckPort
                 this.healthCheckProtocol = endpointGroup.healthCheckProtocol
                 this.thresholdCount = endpointGroup.thresholdCount
-                this.trafficDialPercentage = endpointGroup.trafficDialPercentage.toInt()
+                this.trafficDialPercentage = endpointGroup.trafficDialPercentage.toDouble()
                 this.endpointGroupRegion = endpointGroup.endpointGroupRegion
                 this.endpointConfigurations = getEndpointConfigurations(endpointGroup.endpointDescriptions)
             }
